update our top left avatar on avatar changes
when our primary device updates the avatar, we wait for a sync message
to know about this update.
Once this message comes in, it updates the avatar on the conversation,
but nothing refresh the ActionPanel component.

This commit makes sure we listen for changes on our conversation to
trigger this refresh.