fix: allow legacy config message during linking
We only handle handlethem after the new config message ones, so if there is a
shared config message on the network we will use it for the linking
process instead. The legacy one this commit is about is bound to be
removed completely and will only be used until our users have updated to
the new ones for some time.